put the serve ref on context
